app/requirements.txt
Flask
jsonify
mysql-connector-python